refactor: update import paths for utility components to use the new ui directory structure
