+++ src/components/modules/escrow/ui/dialogs/EscrowDetailDialog.tsx
@@ -14,14 +14,14 @@ import { Escrow } from "@/@types/escrow.entity";
 import { Card, CardContent } from "@/components/ui/card";
 import { cn } from "@/lib/utils";
 import { useFormatUtils } from "@/utils/hook/format.hook";
-import TooltipInfo from "@/components/utils/Tooltip";
+import TooltipInfo from "@/components/utils/ui/Tooltip";
 import { useCopyUtils } from "@/utils/hook/copy.hook";
 import EntityCard from "./cards/EntityCard";
 import FundEscrowDialog from "./FundEscrowDialog";
 import { useEscrowBoundedStore } from "../../store/ui";
 import { useGlobalBoundedStore } from "@/core/store/data";
 import QREscrowDialog from "./QREscrowDialog";
-import LoaderData from "@/components/utils/LoaderData";
+import LoaderData from "@/components/utils/ui/LoaderData";
 import { Badge } from "@/components/ui/badge";
 import useDistributeEarningsEscrowDialogHook from "./hooks/distribute-earnings-escrow-dialog.hook";
 import useChangeStatusEscrowDialogHook from "./hooks/change-status-escrow-dialog.hook";
