+++ src/components/modules/escrow/ui/filters/MyEscrowsFilter.tsx
@@ -6,8 +6,8 @@ import {
   SelectItem,
   SelectTrigger,
 } from "@/components/ui/select";
-import CreateButton from "@/components/utils/Create";
-import Divider from "@/components/utils/Divider";
+import CreateButton from "@/components/utils/ui/Create";
+import Divider from "@/components/utils/ui/Divider";
 import { Search, Trash2 } from "lucide-react";
 
 const MyEscrowsFilter = () => {
