+++ src/components/modules/escrow/ui/tables/MyEscrowsTable.tsx
@@ -20,14 +20,14 @@ import { Input } from "@/components/ui/input";
 import useMyEscrows from "../../hooks/my-escrows.hook";
 import { useFormatUtils } from "@/utils/hook/format.hook";
 import { Escrow } from "@/@types/escrow.entity";
-import NoData from "@/components/utils/NoData";
+import NoData from "@/components/utils/ui/NoData";
 import { useEscrowBoundedStore } from "../../store/ui";
 import EscrowDetailDialog from "../dialogs/EscrowDetailDialog";
 import {
   useGlobalAuthenticationStore,
   useGlobalBoundedStore,
 } from "@/core/store/data";
-import LoaderData from "@/components/utils/LoaderData";
+import LoaderData from "@/components/utils/ui/LoaderData";
 import ExpandableContent from "./expandable/ExpandableContent";
 import SuccessDialog, { SuccessReleaseDialog } from "../dialogs/SuccessDialog";
