+++ src/components/modules/help/ui/pages/Help.tsx
@@ -1,7 +1,7 @@
 import { Bounded } from "@/components/layout/Bounded";
 import HelpAccordion from "@/components/modules/help/ui/utils/HelpAccordion";
 import { Card, CardContent } from "@/components/ui/card";
-import Loader from "@/components/utils/Loader";
+import Loader from "@/components/utils/ui/Loader";
 import { cn } from "@/lib/utils";
 import { useLoaderStore } from "@/store/utilsStore/store";
